Fixed potential circular imports with test functions
- Moved element_exists() and is_scrolled_into_view() to private helper functions in expected_conditions.py
- Updated usages of element_exists() in page object prototypes to just wrap the find_element() calls in a try/except loop, which sets the return values to False if a NoSuchElementException is thrown
