+++ webdriver_test_tools/webdriver/support/expected_conditions.py
@@ -1,13 +1,13 @@
 """Custom expected condition classes"""
 from selenium.common.exceptions import NoSuchElementException
 
-from . import test
 from webdriver_test_tools.common import utils
 
 
-# TODO: add expected conditions for all basic tests that don't have one
 # TODO: add inverse conditions for each? at least for consistency
 # TODO: make class names consistent with selenium's absurd naming conventions
+
+# Expected Condition Classes
 
 class element_to_exist:
     """Custom wait condition for WebdriverWait() that checks if an element
@@ -17,7 +17,7 @@ class element_to_exist:
         self.locator = locator
 
     def __call__(self, driver):
-        return test.element_exists(driver, self.locator)
+        return _element_exists(driver, self.locator)
 
 
 class element_to_not_exist:
@@ -28,7 +28,7 @@ class element_to_not_exist:
         self.locator = locator
 
     def __call__(self, driver):
-        return not test.element_exists(driver, self.locator)
+        return not _element_exists(driver, self.locator)
 
 
 class element_to_be_enabled:
@@ -64,7 +64,7 @@ class element_is_in_view:
 
     def __call__(self, driver):
         element = driver.find_element(*self.locator)
-        return test.is_scrolled_into_view(driver, element, self.fully_in_view)
+        return _is_scrolled_into_view(driver, element, self.fully_in_view)
 
 
 class base_url_to_be:
@@ -83,3 +83,55 @@ class base_url_to_be:
         base_url = utils.get_base_url(driver.current_url)
         return self.url == base_url
 
+
+# Helper Methods
+
+def _is_scrolled_into_view(driver, element, fully_in_view=True):
+    """Returns True if the element is scrolled into view, False otherwise
+
+    Currently, Selenium doesn't offer a means of getting an element's location relative
+    to the viewport, so using JavaScript to determine whether the element is visible
+    within the viewport.
+
+    :param driver: Selenium WebDriver object
+    :param element: WebElement for the element to check
+    :param fully_in_view: (Default = True) If True, check that the element is fully in
+        view and not cut off. If False, check that it's at least partially in view
+
+    :return: True if the element is scrolled into view, False otherwise
+    """
+    # the JavaScript used to check if the element is in view.
+    script_string = '''
+    return function(el, strict) {
+        var rect = el.getBoundingClientRect();
+        var elemTop = rect.top;
+        var elemBottom = rect.bottom;
+
+        if (strict)
+            var isVisible = (elemTop >= 0) && (elemBottom <= window.innerHeight);
+        else
+            isVisible = elemTop < window.innerHeight && elemBottom >= 0;
+        return isVisible;
+    }(arguments[0],arguments[1])
+    '''
+    return driver.execute_script(script_string, element, fully_in_view)
+
+
+def _element_exists(driver, element_locator):
+    """Returns True if the element exists, False if not
+
+    This function is just a wrapper that catches the NoSuchElementException thrown by
+    driver.find_element() and returns a boolean based on whether the exception occurred.
+    Used for test assertions.
+
+    :param driver: Selenium WebDriver object
+    :param element_locator: Tuple in the format (by,selector) used to locate target
+
+    :return: True if the element exists, False if not
+    """
+    exists = True
+    try:
+        driver.find_element(*element_locator)
+    except NoSuchElementException:
+        exists = False
+    return exists
